Correction bug - Implémentation fonctionnalités Stagiaire corrigée

+++ src/main/java/com/hrizzon2/demotest/dto/stagiaire/StagiaireDTO.java
@@ -16,9 +16,9 @@ public class StagiaireDTO {
     private String firstName;
     private String email;
     private Number phone;
-    private Long villeId;
+    private Integer villeId;
     private String villeNom;
-    private List<Long> dossiersIds;
+    private List<Integer> dossiersIds;
 
 }
